+++ src/App.scss
@@ -48,6 +48,7 @@ $breakpoint-small: 460px;
 
 @import "scss/app-header";
 @import "scss/app-footer";
+@import "scss/npc-options";
 @import "scss/npc-card";
 @import "scss/npc-stat";
 @import "scss/generate-npc-button";
