+++ src/components/npc-card.tsx
@@ -1,271 +1,14 @@
 import React, { Component } from 'react';
-import { SimpleRange } from '../interfaces/simple-range';
 import { Stat } from '../interfaces/stat';
-import { Alignment } from '../interfaces/alignment';
 import { gods } from '../consts/gods';
-
-class RaceInfo {
-	name: string;
-	age: SimpleRange = { min: 0, max: 0 };
-	height: SimpleRange = { min: 0, max: 0 };
-	weight: SimpleRange = { min: 0, max: 0 };
-	names: Map<string, string[]> = new Map<string, string[]>();
-
-	constructor(name: string, minAge: number, maxAge: number, minHeight: number, maxHeight: number, minWeight: number, maxWeight: number) {
-		this.setNames(name.toLowerCase());
-		this.name = name;
-		this.age.min = minAge;
-		this.age.max = maxAge;
-		this.height.min = minHeight;
-		this.height.max = maxHeight;
-		this.weight.min = minWeight;
-		this.weight.max = maxWeight;
-	}
-
-	setNames(name: string): void {
-		switch (name) {
-			case 'dragonborn':
-				this.names.set('male', ['Arjhan', 'Balasar', 'Bharash', 'Donaar', 'Ghesh', 'Heskan', 'Kriv', 'Medrash', 'Mehen', 'Nadarr', 'Pandjed', 'Patrin', 'Rhogar', 'Shamash', 'Shedinn', 'Tarhun', 'Torinn']);
-				this.names.set('female', ['Akra', 'Biri', 'Daar', 'Farideh', 'Harann', 'Havilar', 'Jheri', 'Kava', 'Korinn', 'Mishann', 'Nala', 'Perra', 'Raiann', 'Sora', 'Surina', 'Thava', 'Uadjit']);
-				this.names.set('surname', ['Clethtinthiallor', 'Daardendrian', 'Delmirev', 'Drachedandion', 'Fenkenkabradon', 'Kepeshkmolik', 'Kerrhylon', 'Kimbatuul', 'Linxakasendalor', 'Myastan', 'Nemmonis', 'Norixius', 'Ophinshtalajiir', 'Prexijandilin', 'Shestendeliath', 'Turnuroth', 'Verthisathurgiesh', 'Yarjerit']);
-				break;
-			case 'dwarf':
-				this.names.set('male', ['Adrik', 'Alberich', 'Baern', 'Barendd', 'Brottor', 'Bruenor', 'Dain', 'Darrak', 'Delg', 'Eberk', 'Einkil', 'Fargrim', 'Flint', 'Gardain', 'Harbek', 'Kildrak', 'Morgran', 'Orsik', 'Oskar', 'Rangrim', 'Rurik', 'Taklinn', 'Thoradin', 'Thorin', 'Tordek', 'Traubon', 'Travok', 'Ulfgar', 'Veit', 'Vondal']);
-				this.names.set('female', ['Amber', 'Artin', 'Audhild', 'Bardryn', 'Dagnal', 'Diesa', 'Eldeth', 'Falkrunn', 'Finellen', 'Gunnloda', 'Gurdis', 'Helja', 'Hlin', 'Kathra', 'Kristryd', 'Ilde', 'Liftrasa', 'Mardred', 'Riswynn', 'Sannl', 'Torbera', 'Torgga', 'Vistra']);
-				this.names.set('surname', ['Balderk', 'Battlehammer', 'Brawnanvil', 'Dankil', 'Fireforge', 'Frostbeard', 'Gorunn', 'Holderhek', 'Ironfist', 'Loderr', 'Lutgehr', 'Rumnaheim', 'Strakeln', 'Torunn', 'Ungart']);
-				break;
-			case 'elf':
-				this.names.set('male', ['Adran', 'Aelar', 'Aramil', 'Arannis', 'Aust', 'Beiro', 'Berrian', 'Carric', 'Enialis', 'Erdan', 'Erevan', 'Galinndan', 'Hadarai', 'Heian', 'Himo', 'Immeral', 'Ivellios', 'Laucian', 'Mindartis', 'Paelias', 'Peren', 'Quarion', 'Riardon', 'Rolen', 'Soveliss', 'Thamior', 'Tharivol', 'Theren', 'Varis', 'Aseir', 'Bardeid', 'Haseid', 'Khemed', 'Mehmen', 'Sudeiman', 'Zasheir', 'Darvin', 'Dorn', 'Evendur', 'Gorstag', 'Grim', 'Helm', 'Malark', 'Morn', 'Randal', 'Stedd', 'Bor', 'Fodel', 'Glar', 'Grigor', 'Igan', 'Ivor', 'Kosef', 'Mival', 'Orel', 'Pavel', 'Sergor', 'Ander', 'Blath', 'Bran', 'Frath', 'Geth', 'Lander', 'Luth', 'Malcer', 'Stor', 'Taman', 'Urth Aoth', 'Bareris', 'Ehput-Ki', 'Kethoth', 'Mumed', 'Ramas', 'So-Kehur', 'Thazar-De', 'Urhur', 'Borivik', 'Faurgar', 'Jandar', 'Kanithar', 'Madislak', 'Ralmevik', 'Shaumar', 'Vladislak', 'An', 'Chen', 'Chi', 'Fai', 'Jiang', 'Jun', 'Lian', 'Long', 'Meng', 'On', 'Shan', 'Shui', 'Wen', 'Anton', 'Diero', 'Marcon', 'Pieron', 'Rimardo', 'Romero', 'Salazar', 'Umbero']);
-				this.names.set('female', ['Adrie', 'Althaea', 'Anastrianna', 'Andraste', 'Antinua', 'Bethrynna', 'Birel', 'Caelynn', 'Drusilia', 'Enna', 'Felosial', 'Ielenia', 'Jelenneth', 'Keyleth', 'Leshanna', 'Lia', 'Meriele', 'Mialee', 'Naivara', 'Quelenna', 'Quillathe', 'Sariel', 'Shanairra', 'Shava', 'Silaqui', 'Theirastra', 'Thia', 'Vadania', 'Valanthe', 'Xanaphia']);
-				this.names.set('surname', ['Amakiir (Gemflower)', 'Amastacia (Starflower)', 'Galanodel (Moonwhisper)', 'Holimion (Diamonddew)', 'Ilphelkiir (Gemblossom)', 'Liadon (Silverfrond)', 'Meliamne (Oakenheel)', 'Naïlo (Nightbreeze)', 'Siannodel (Moonbrook)', 'Xiloscient (Goldpetal)']);
-				break;
-			case 'gnome':
-				this.names.set('male', ['Alston', 'Alvyn', 'Boddynock', 'Brocc', 'Burgell', 'Dimble', 'Eldon', 'Erky', 'Fonkin', 'Frug', 'Gerbo', 'Gimble', 'Glim', 'Jebeddo', 'Kellen', 'Namfoodle', 'Orryn', 'Roondar', 'Seebo', 'Sindri', 'Warryn', 'Wrenn', 'Zook']);
-				this.names.set('female', ['Bimpnottin', 'Breena', 'Caramip', 'Carlin', 'Donella', 'Duvamil', 'Ella', 'Ellyjobell', 'Ellywick', 'Lilli', 'Loopmottin', 'Lorilla', 'Mardnab', 'Nissa', 'Nyx', 'Oda', 'Orla', 'Roywyn', 'Shamil', 'Tana', 'Waywocket', 'Zanna']);
-				this.names.set('surname', ['Beren', 'Daergel', 'Folkor', 'Garrick', 'Nackle', 'Murnig', 'Ningel', 'Raulnor', 'Scheppen', 'Timbers', 'Turen']);
-				break;
-			case 'half-elf':
-				// combine elf and human
-				this.names.set('male', ['Adran', 'Aelar', 'Aramil', 'Arannis', 'Aust', 'Beiro', 'Berrian', 'Carric', 'Enialis', 'Erdan', 'Erevan', 'Galinndan', 'Hadarai', 'Heian', 'Himo', 'Immeral', 'Ivellios', 'Laucian', 'Mindartis', 'Paelias', 'Peren', 'Quarion', 'Riardon', 'Rolen', 'Soveliss', 'Thamior', 'Tharivol', 'Theren', 'Varis']);
-				this.names.set('female', ['Adrie', 'Althaea', 'Anastrianna', 'Andraste', 'Antinua', 'Bethrynna', 'Birel', 'Caelynn', 'Drusilia', 'Enna', 'Felosial', 'Ielenia', 'Jelenneth', 'Keyleth', 'Leshanna', 'Lia', 'Meriele', 'Mialee', 'Naivara', 'Quelenna', 'Quillathe', 'Sariel', 'Shanairra', 'Shava', 'Silaqui', 'Theirastra', 'Thia', 'Vadania', 'Valanthe', 'Xanaphia', 'Atala', 'Ceidil', 'Hama', 'Jasmal', 'Meilil', 'Seipora', 'Yasheira', 'Zasheida', 'Arveene', 'Esvele', 'Jhessail', 'Kerri', 'Lureene', 'Miri', 'Rowan', 'Shandri', 'Tessele', 'Alethra', 'Kara', 'Katernin', 'Mara', 'Natali', 'Olma', 'Tana', 'Zora', 'Amafrey', 'Betha', 'Cefrey', 'Kethra', 'Mara', 'Olga', 'Silifrey', 'Westra', 'Arizima', 'Chathi', 'Nephis', 'Nulara', 'Murithi', 'Sefris', 'Thola', 'Umara', 'Zolis', 'Fyevarra', 'Hulmarra', 'Immith', 'Imzel', 'Navarra', 'Shevarra', 'Tammith', 'Yuldra', 'Bai', 'Chao', 'Jia', 'Lei', 'Mei', 'Qiao', 'Shui', 'Tai', 'Balama', 'Dona', 'Faila', 'Jalana', 'Luisa', 'Marta', 'Quara', 'Selise', 'Vonda']);
-				this.names.set('surname', ['Amakiir (Gemflower)', 'Amastacia (Starflower)', 'Galanodel (Moonwhisper)', 'Holimion (Diamonddew)', 'Ilphelkiir (Gemblossom)', 'Liadon (Silverfrond)', 'Meliamne (Oakenheel)', 'Naïlo (Nightbreeze)', 'Siannodel (Moonbrook)', 'Xiloscient (Goldpetal)', 'Basha', 'Dumein', 'Jassan', 'Khalid', 'Mostana', 'Pashar', 'Rein', 'Amblecrown', 'Buckman', 'Dundragon', 'Evenwood', 'Greycastle', 'Tallstag', 'Bersk', 'Chernin', 'Dotsk', 'Kulenov', 'Marsk', 'Nemetsk', 'Shemov', 'Starag', 'Brightwood', 'Helder', 'Hornraven', 'Lackman', 'Stormwind', 'Windrivver', 'Ankhalab', 'Anskuld', 'Fezim', 'Hahpet', 'Nathandem', 'Sepret', 'Uuthrakt', 'Chergoba', 'Dyernina', 'Iltazyara', 'Murnyethara', 'Stayanoga', 'Ulmokina', 'Chien', 'Huang', 'Kao', 'Kung', 'Lao', 'Ling', 'Mei', 'Pin', 'Shin', 'Sum', 'Tan', 'Wan', 'Agosto', 'Astorio', 'Calabra', 'Domine', 'Falone', 'Marivaldi', 'Pisacar', 'Ramondo']);
-				break;
-			case 'half-orc':
-				// combine human and orc
-				this.names.set('male', ['Dench', 'Feng', 'Gell', 'Henk', 'Holg', 'Imsh', 'Keth', 'Krusk', 'Mhurren', 'Ront', 'Shump', 'Thokk']);
-				this.names.set('female', ['Baggi', 'Emen', 'Engong', 'Kansif', 'Myev', 'Neega', 'Ovak', 'Ownka', 'Shautha', 'Sutha', 'Vola', 'Volen', 'Yevelda']);
-				// no last names
-				this.names.set('surname', ['']);
-				break;
-			case 'halfling':
-				this.names.set('male', ['Alton', 'Ander', 'Cade', 'Corrin', 'Eldon', 'Errich', 'Finnan', 'Garret', 'Lindal', 'Lyle', 'Merric', 'Milo', 'Osborn', 'Perrin', 'Reed', 'Roscoe', 'Wellby']);
-				this.names.set('female', ['Andry', 'Bree', 'Callie', 'Cora', 'Euphemia', 'Jillian', 'Kithri', 'Lavinia', 'Lidda', 'Merla', 'Nedda', 'Paela', 'Portia', 'Seraphina', 'Shaena', 'Trym', 'Vani', 'Verna']);
-				this.names.set('surname', ['Brushgather', 'Goodbarrel', 'Greenbottle', 'High-hill', 'Hilltopple', 'Leagallow', 'Tealeaf', 'Thorngage', 'Tosscobble', 'Underbough']);
-				break;
-			case 'human':
-				this.names.set('male', ['Aseir', 'Bardeid', 'Haseid', 'Khemed', 'Mehmen', 'Sudeiman', 'Zasheir', 'Darvin', 'Dorn', 'Evendur', 'Gorstag', 'Grim', 'Helm', 'Malark', 'Morn', 'Randal', 'Stedd', 'Bor', 'Fodel', 'Glar', 'Grigor', 'Igan', 'Ivor', 'Kosef', 'Mival', 'Orel', 'Pavel', 'Sergor', 'Ander', 'Blath', 'Bran', 'Frath', 'Geth', 'Lander', 'Luth', 'Malcer', 'Stor', 'Taman', 'Urth Aoth', 'Bareris', 'Ehput-Ki', 'Kethoth', 'Mumed', 'Ramas', 'So-Kehur', 'Thazar-De', 'Urhur', 'Borivik', 'Faurgar', 'Jandar', 'Kanithar', 'Madislak', 'Ralmevik', 'Shaumar', 'Vladislak', 'An', 'Chen', 'Chi', 'Fai', 'Jiang', 'Jun', 'Lian', 'Long', 'Meng', 'On', 'Shan', 'Shui', 'Wen', 'Anton', 'Diero', 'Marcon', 'Pieron', 'Rimardo', 'Romero', 'Salazar', 'Umbero']);
-				this.names.set('female', ['Atala', 'Ceidil', 'Hama', 'Jasmal', 'Meilil', 'Seipora', 'Yasheira', 'Zasheida', 'Arveene', 'Esvele', 'Jhessail', 'Kerri', 'Lureene', 'Miri', 'Rowan', 'Shandri', 'Tessele', 'Alethra', 'Kara', 'Katernin', 'Mara', 'Natali', 'Olma', 'Tana', 'Zora', 'Amafrey', 'Betha', 'Cefrey', 'Kethra', 'Mara', 'Olga', 'Silifrey', 'Westra', 'Arizima', 'Chathi', 'Nephis', 'Nulara', 'Murithi', 'Sefris', 'Thola', 'Umara', 'Zolis', 'Fyevarra', 'Hulmarra', 'Immith', 'Imzel', 'Navarra', 'Shevarra', 'Tammith', 'Yuldra', 'Bai', 'Chao', 'Jia', 'Lei', 'Mei', 'Qiao', 'Shui', 'Tai', 'Balama', 'Dona', 'Faila', 'Jalana', 'Luisa', 'Marta', 'Quara', 'Selise', 'Vonda']);
-				this.names.set('surname', ['Basha', 'Dumein', 'Jassan', 'Khalid', 'Mostana', 'Pashar', 'Rein', 'Amblecrown', 'Buckman', 'Dundragon', 'Evenwood', 'Greycastle', 'Tallstag', 'Bersk', 'Chernin', 'Dotsk', 'Kulenov', 'Marsk', 'Nemetsk', 'Shemov', 'Starag', 'Brightwood', 'Helder', 'Hornraven', 'Lackman', 'Stormwind', 'Windrivver', 'Ankhalab', 'Anskuld', 'Fezim', 'Hahpet', 'Nathandem', 'Sepret', 'Uuthrakt', 'Chergoba', 'Dyernina', 'Iltazyara', 'Murnyethara', 'Stayanoga', 'Ulmokina', 'Chien', 'Huang', 'Kao', 'Kung', 'Lao', 'Ling', 'Mei', 'Pin', 'Shin', 'Sum', 'Tan', 'Wan', 'Agosto', 'Astorio', 'Calabra', 'Domine', 'Falone', 'Marivaldi', 'Pisacar', 'Ramondo']);
-				break;
-			case 'tiefling':
-				this.names.set('male', ['Akmenos', 'Amnon', 'Barakas', 'Damakos', 'Ekemon', 'Iados', 'Kairon', 'Leucis', 'Melech', 'Mordai', 'Morthos', 'Pelaios', 'Skamos', 'Therai']);
-				this.names.set('female', ['Akta', 'Anakis', 'Bryseis', 'Criella', 'Damaia', 'Ea', 'Kallista', 'Lerissa', 'Makaria', 'Nemeia', 'Orianna', 'Phelaia', 'Rieta']);
-				// virtue names
-				this.names.set('surname', ['Art', 'Carrion', 'Chant', 'Creed', 'Despair', 'Excellence', 'Fear', 'Glory', 'Hope', 'Ideal', 'Music', 'Nowhere', 'Open', 'Poetry', 'Quest', 'Random', 'Reverence', 'Sorrow', 'Temerity', 'Torment', 'Weary']);
-				break;
-			default:
-				this.names.set('male', ['Josh', 'Mike', 'Pat', 'Justin', 'Hector The Well Endowed', 'Marrrrrr']);
-				this.names.set('female', ['Jess', 'Shannon', 'Mal', 'Riley']);
-				this.names.set('surname', ['Calderon', 'Batman', 'Graveler', 'Peepee', 'Poopoo']);
-				break;
-		}
-	}
-}
-
+import { races } from '../consts/races';
+import { alignments } from '../consts/alignments';
+import { adjectives, traits, religiousAdjective, socioeconomic, traits2, quirks } from '../consts/npc-bio';
+import { randomNumber, getRandomNumberStandardDist, getRandomMapKey } from '../utility/functions';
 
 const baseUrl: string = 'https://www.dnd5eapi.co/api/';
 const abilities: string[] = ['strength', 'dexterity', 'constitution', 'intelligence', 'wisdom', 'charisma'];
 const classes: string[] = ['barbarian', 'bard', 'cleric', 'druid', 'fighter', 'monk', 'paladin', 'ranger', 'rogue', 'sorcerer', 'warlock', 'wizard'];
-const alignments: Map<string, Alignment> = new Map<string, Alignment>([
-	['chaotic-evil', { alias: 'chaotic-evil', code: 'CE', name: 'Chaotic Evil' }],
-	['chaotic-neutral', { alias: 'chaotic-neutral', code: 'CN', name: 'Chaotic Neutral' }],
-	['chaotic-good', { alias: 'chaotic-good', code: 'CG', name: 'Chaotic Good' }],
-	['lawful-evil', { alias: 'lawful-evil', code: 'LE', name: 'Lawful Evil' }],
-	['lawful-neutral', { alias: 'lawful-neutral', code: 'LN', name: 'Lawful Neutral' }],
-	['lawful-good', { alias: 'lawful-good', code: 'LG', name: 'Lawful Good' }],
-	['neutral-evil', { alias: 'neutral-evil', code: 'NE', name: 'Neutral Evil' }],
-	['neutral', { alias: 'neutral', code: 'N', name: 'Neutral' }],
-	['neutral-good', { alias: 'neutral-good', code: 'NG', name: 'Neutral Good' }],
-]);
-const races: Map<string, RaceInfo> = new Map<string, RaceInfo>([
-	['dragonborn', new RaceInfo('Dragonborn', 15, 80, 78, 90, 215, 300)],
-	['dwarf', new RaceInfo('Dwarf', 50, 350, 48, 60, 120, 200)],
-	['elf', new RaceInfo('Elf', 100, 750, 54, 78, 125, 215)],
-	['gnome', new RaceInfo('Gnome', 40, 480, 32, 48, 40, 45)],
-	['half-elf', new RaceInfo('Half-Elf', 20, 190, 58, 74, 140, 220)],
-	['half-orc', new RaceInfo('Half-Orc', 14, 75, 60, 80, 150, 250)],
-	['halfling', new RaceInfo('Halfling', 20, 250, 32, 42, 35, 45)],
-	['human', new RaceInfo('Human', 18, 80, 60, 78, 130, 225)],
-	['tiefling', new RaceInfo('Tiefling', 18, 90, 60, 78, 130, 225)]
-]);
-// const gods: Map<string, string[]> = new Map<string, string[]>([
-// 	['CE', ['Beshaba, goddess of misfortune', 'Cyric, god of lies', 'Malar, god of the hunt', 'Talona, goddess of disease and poison', 'Talos, god of storms', 'Umberlee, goddess of the sea']],
-// 	['CN', ['Leira, goddess of illusion', 'Mask, god of thieves']],
-// 	['CG', ['Lliira, goddess of joy', 'Selûne, goddess of the moon', 'Sune, goddess of love and beauty', 'Tymora, goddess of good fortune']],
-// 	['LE', ['Bane, god of tyranny', 'Loviatar, goddess of pain']],
-// 	['LN', ['Azuth, god of wizards', 'Helm, god of protection', 'Kelemvor, god of the dead', 'Savras, god of divination and fate']],
-// 	['LG', ['Ilmater, god of endurance', 'Torm, god of courage and self-sacrifice', 'Tyr, god of justice']],
-// 	['NE', ['Auril, goddess of winter', 'Bhaal, god of murder', 'Myrkul, god of death', 'Shar, goddess of darkness and loss']],
-// 	['N', ['Gond, god of craft', 'Oghma, god of knowledge', 'Silvanus, god of wild nature', 'Tempus, god of war', 'Waukeen, goddess of trade']],
-// 	['NG', ['Chauntea, goddess of agriculture', 'Deneir, god of writing', 'Eldath, goddess of peace', 'Lathander, god of birth and renewal', 'Mielikki, goddess of forests', 'Milil, god of poetry and song', 'Mystra, goddess of magic']]
-// ]);
-
-/*
-*
-* These constants are used to generate the NCp's bio
-*
-*/
-
-// General adjective to describe personality
-const adjectives: string[] = ['fiesty', 'stubborn', 'timid', 'sneaky', 'aggressive', 'horny', 'punctual'];
-
-// Things they do
-const traits: string[] = [
-	'who runs their family\'s business.',
-	'who collects toads in their free time.',
-	'with more clothing options than common sense.',
-	'that loves their dog more than anything.',
-	'with a love-hate relationship with cooking.'
-]
-
-// Ways they behave
-const traits2: string[] = [
-	'They are a skilled combatant and are proud of it.',
-	'They are not a fan of small talk.',
-	'They look for the good in everyone they meet.',
-	'They manage to find something wrong with everything.'
-]
-
-// Modifies how strong their faith is
-const religiousAdjective: string[] = [
-	'are a devout',
-	'are a zealous',
-	'pretend to be a',
-	'are a nominal',
-	'are a bregrudging'
-];
-
-// Describes how much wealth they have
-const socioeconomic: string[] = [
-	'are positively abreast with currency.',
-	'are poor as fuck.',
-	'were born rich but are terrible with money.',
-	'live comfortably.',
-	'though they have little money, they are content.'
-];
-
-// A quirk they have to give them some el oh el ha ha
-const quirks: string[] = [
-	'They are an aspiring matress salesman.',
-	'Their favorite thing to do is eat pinecones.',
-	'They believe they are a God stuck in mortal form.',
-	'They don\'t like cheese unless it is frozen solid.',
-	'They like to watch.',
-	'They always wear an eye patch but often change which eye. No one seems to notice.',
-	'Their nickname is "Jabonicus" but refuses to elaborate on why.',
-	'They have died three times... so far.',
-	'They were raised by wolves and are searching for their two brothers, Vin and Paul.',
-	'They are an accomplished boar racer.',
-	'They have a wooden pegleg that is filled with a secret.',
-	'When they hear a special word they instantly fall asleep.'
-];
-
-/*
-*
-* Utility
-*
-*/
-
-// Answer by joshuakcockrell: https://stackoverflow.com/questions/25582882/javascript-math-random-normal-distribution-gaussian-bell-curve
-function getRandomNumberStandardDist(min: number, max: number, skew: number) {
-	let u = 0, v = 0;
-	while (u === 0) u = Math.random() //Converting [0,1) to (0,1)
-	while (v === 0) v = Math.random()
-	let num = Math.sqrt(-2.0 * Math.log(u)) * Math.cos(2.0 * Math.PI * v)
-
-	num = num / 10.0 + 0.5 // Translate to 0 -> 1
-	if (num > 1 || num < 0)
-		num = getRandomNumberStandardDist(min, max, skew) // resample between 0 and 1 if out of range
-
-	else {
-		num = Math.pow(num, skew) // Skew
-		num *= max - min // Stretch to fill range
-		num += min // offset to min
-	}
-	return Math.floor(num);
-}
-
-function randomNumber(min: number, max: number) {
-	return Math.floor(Math.random() * (max - min) + min);
-}
-
-function getRandomMapKey(map: Map<any, any>): string {
-	const items: string[] = Array.from(map.keys());
-	return items[Math.floor(randomNumber(0, items.length))];
-}
-
-function capitalize(word: string): string {
-	return word.split('').map((letter, i) => i ? letter.toLowerCase() : letter.toUpperCase()).join('');
-}
-
-// Functions below this comment should probably be in the component
-
-function getNpcGender(): string {
-	return Math.floor(Math.random() * 2) === 0 ? 'Male' : 'Female';
-}
-
-function getArmorClass(score: number): number {
-	return 10 + getModifier(score);
-}
-
-function getHitpoints(hitDie: number, score: number, level?: number): number {
-	let hp = hitDie + getModifier(score),
-		rollsLeft = level || 0;
-
-	while (rollsLeft > 1) {
-		// TODO YeahThe modifier should adjust as con increases but it doesn't right now
-		hp += (randomNumber(1, hitDie) + getModifier(score));
-		rollsLeft--;
-	}
-
-	return hp;
-}
-
-function getHeight(min: number, max: number, skew: number): string {
-	const height = getRandomNumberStandardDist(min, max, skew),
-		feet = Math.floor(height / 12),
-		inches = height % 12;
-
-	return feet + "' " + inches + '"';
-}
-
-function getNpcName(race: string, gender: string): string {
-	const rand1 = Math.floor(Math.random() * races.get(race)!.names.get(gender.toLowerCase())!.length),
-		rand2 = Math.floor(Math.random() * races.get(race)!.names.get('surname')!.length);
-
-	let last = races.get(race)!.names.get('surname')![rand2],
-		first = races.get(race)!.names.get(gender.toLowerCase())![rand1];
-
-
-	let returnName = first;
-	if (last) { returnName = returnName + " " + last; }
-
-	return returnName;
-}
-
-function getModifier(score: number): number {
-	return Math.floor((score - 10) / 2);
-}
 
 function NpcLineItem(props: any) {
 	return (
@@ -332,6 +75,53 @@ export class NpcCard extends Component<any, any> {
 		};
 	}
 
+	getNpcGender(): string {
+		return Math.floor(Math.random() * 2) === 0 ? 'Male' : 'Female';
+	}
+
+	getArmorClass(score: number): number {
+		return 10 + this.getModifier(score);
+	}
+
+	getHitpoints(hitDie: number, score: number, level?: number): number {
+		let hp = hitDie + this.getModifier(score),
+			rollsLeft = level || 0;
+
+		while (rollsLeft > 1) {
+			// TODO The modifier should adjust as con increases but it doesn't right now
+			hp += (randomNumber(1, hitDie) + this.getModifier(score));
+			rollsLeft--;
+		}
+
+		return hp;
+	}
+
+	getHeight(min: number, max: number, skew: number): string {
+		const height = getRandomNumberStandardDist(min, max, skew),
+			feet = Math.floor(height / 12),
+			inches = height % 12;
+
+		return feet + "' " + inches + '"';
+	}
+
+	getNpcName(race: string, gender: string): string {
+		const rand1 = Math.floor(Math.random() * races.get(race)!.names.get(gender.toLowerCase())!.length),
+			rand2 = Math.floor(Math.random() * races.get(race)!.names.get('surname')!.length);
+
+		let last = races.get(race)!.names.get('surname')![rand2],
+			first = races.get(race)!.names.get(gender.toLowerCase())![rand1];
+
+
+		let returnName = first;
+		if (last) { returnName = returnName + " " + last; }
+
+		return returnName;
+	}
+
+	getModifier(score: number): number {
+		return Math.floor((score - 10) / 2);
+	}
+
 	generateStats(hitDie: number, abilityBonuses: any, numOfAbi: number, level: number): void {
 		let stats: { [key: string]: {} } = {},
 			ac: number = 10,
@@ -361,16 +151,17 @@ export class NpcCard extends Component<any, any> {
 		}
 
 		abilities.forEach((key: string) => {
-			let score: number = this.generateStat();
+			let score: number = this.generateStat(),
+				keyTrunc: string = key.substring(0, 3);
 
-			if (bonuses && bonuses.has(key.substring(0, 3))) {
-				//@ts-ignore I check if the key exists above
-				score += bonuses.get(key.substring(0, 3));
+			if (bonuses && bonuses.has(keyTrunc)) {
+				// @ts-ignore
+				score += bonuses.get(keyTrunc);
 			}
 
-			if (key === 'dexterity') { ac = getArmorClass(score); }
-			if (key === 'constitution') { hitpoints = getHitpoints(hitDie, score, level); }
-			stats[key] = { score: score, modifier: getModifier(score) };
+			if (key === 'dexterity') { ac = this.getArmorClass(score); }
+			if (key === 'constitution') { hitpoints = this.getHitpoints(hitDie, score, level); }
+			stats[key] = { score: score, modifier: this.getModifier(score) };
 		});
 
 		this.setState({
@@ -400,10 +191,11 @@ export class NpcCard extends Component<any, any> {
 	}
 
 	generateDescription = () => {
+		const npc = this.state;
 		return (
 			<>
-				<p>{this.state.npc.name} is a {adjectives[randomNumber(0, adjectives.length)]} {this.state.npc.race.name} {traits[randomNumber(0, traits.length)]}</p>
-				<p>They {religiousAdjective[randomNumber(0, religiousAdjective.length)]} follower of {gods.get(this.state.npc.alignment.abbreviation)![0]} and {socioeconomic[randomNumber(0, socioeconomic.length)]} {traits2[randomNumber(0, traits2.length)]}</p>
+				<p>{npc.name} is a {adjectives[randomNumber(0, adjectives.length)]} {npc.race.name} {traits[randomNumber(0, traits.length)]}</p>
+				<p>They {religiousAdjective[randomNumber(0, religiousAdjective.length)]} follower of {gods.get(npc.alignment.abbreviation)![0]} and {socioeconomic[randomNumber(0, socioeconomic.length)]} {traits2[randomNumber(0, traits2.length)]}</p>
 				<p>{quirks[randomNumber(0, quirks.length)]}</p>
 			</>)
 	}
@@ -421,10 +213,10 @@ export class NpcCard extends Component<any, any> {
 
 		Promise.all([npcRace, npcClass, npcLevel, npcAlignment]).then((r) => {
 			this.generateStats(r[1].class.hit_die, r[0].race.ability_bonuses, r[2].level.ability_score_bonuses, level);
-			const gender = getNpcGender();
+			const gender = this.getNpcGender();
 
 			this.setState({
-				name: getNpcName(r[0].race.index, gender),
+				name: this.getNpcName(r[0].race.index, gender),
 				gender: gender,
 				alignment: r[3].alignment
 			});
@@ -488,7 +280,6 @@ export class NpcCard extends Component<any, any> {
 							<ol className="npc-card__info">
 								{this.renderLineItem(npc.race.name)}
 								{this.renderLineItem(npc.class.name)}
-
 							</ol>
 						</div>
 						<ol className="npc-card__sturdiness">
@@ -503,14 +294,12 @@ export class NpcCard extends Component<any, any> {
 					<div className="npc-card__bio">
 						<ol className="npc-card__attributes">
 							<li><strong>Age</strong>{getRandomNumberStandardDist(races.get(npc.race.index)!.age.min, races.get(npc.race.index)!.age.max, 1)} years</li>
-							<li><strong>Height</strong>{getHeight(races.get(npc.race.index)!.height.min, races.get(npc.race.index)!.height.max, 1)}</li>
+							<li><strong>Height</strong>{this.getHeight(races.get(npc.race.index)!.height.min, races.get(npc.race.index)!.height.max, 1)}</li>
 							<li><strong>Weight</strong>{getRandomNumberStandardDist(races.get(npc.race.index)!.weight.min, races.get(npc.race.index)!.weight.max, 1)} lbs</li>
 							<li><strong>Alignment</strong>{npc.alignment.name}</li>
 						</ol>
 						<div className="npc-card__description">
-							<p>{npc.name} is a {adjectives[randomNumber(0, adjectives.length)]} {npc.race.name} {traits[randomNumber(0, traits.length)]}</p>
-							<p>They {religiousAdjective[randomNumber(0, religiousAdjective.length)]} follower of {gods.get(npc.alignment.abbreviation)![0]} and {socioeconomic[randomNumber(0, socioeconomic.length)]} {traits2[randomNumber(0, traits2.length)]}</p>
-							<p>{quirks[randomNumber(0, quirks.length)]}</p>
+							{this.generateDescription()}
 						</div>
 					</div>
 					<ul className="npc-card__ability-scores" >
