+++ src/consts/npc-bio.ts
@@ -1,0 +1,59 @@
+/*
+*
+* These constants are used to generate the Nc'cs bio
+*
+*/
+
+// General adjective to describe personality
+export const adjectives: string[] = ['fiesty', 'stubborn', 'timid', 'sneaky', 'aggressive', 'horny', 'punctual'];
+
+// Things they do
+export const traits: string[] = [
+	'who runs their family\'s business.',
+	'who collects toads in their free time.',
+	'with more clothing options than common sense.',
+	'that loves their dog more than anything.',
+	'with a love-hate relationship with cooking.'
+]
+
+// Ways they behave
+export const traits2: string[] = [
+	'They are a skilled combatant and are proud of it.',
+	'They are not a fan of small talk.',
+	'They look for the good in everyone they meet.',
+	'They manage to find something wrong with everything.'
+]
+
+// Modifies how strong their faith is
+export const religiousAdjective: string[] = [
+	'are a devout',
+	'are a zealous',
+	'pretend to be a',
+	'are a nominal',
+	'are a bregrudging'
+];
+
+// Describes how much wealth they have
+export const socioeconomic: string[] = [
+	'are positively abreast with currency.',
+	'are poor as fuck.',
+	'were born rich but are terrible with money.',
+	'live comfortably.',
+	'though they have little money, they are content.'
+];
+
+// A quirk they have to give them some el oh el ha ha
+export const quirks: string[] = [
+	'They are an aspiring matress salesman.',
+	'Their favorite thing to do is eat pinecones.',
+	'They believe they are a God stuck in mortal form.',
+	'They don\'t like cheese unless it is frozen solid.',
+	'They like to watch.',
+	'They always wear an eye patch but often change which eye. No one seems to notice.',
+	'Their nickname is "Jabonicus" but refuses to elaborate on why.',
+	'They have died three times... so far.',
+	'They were raised by wolves and are searching for their two brothers, Vin and Paul.',
+	'They are an accomplished boar racer.',
+	'They have a wooden pegleg that is filled with a secret.',
+	'When they hear a special word they instantly fall asleep.'
+];
